Integrated Dropbox functionality
Made minor changes to the code that would allow for document saving; whenever
the document would save locally in `todo-behaviour`, it would also attempt
to be saved to Dropbox (fails if no Dropbox linkage).  The requests to save
to Dropbox are handled by `todo-controller`, just as saving to the local file
system is.